+++ R/subtitle_maker_templates.R
@@ -28,7 +28,7 @@
 #'
 #' @examples
 #' set.seed(123)
-#'
+#' 
 #' # subtitle for *t*-statistic with Cohen's *d* as effect size
 #' ggstatsplot:::subtitle_template_1(
 #'   statistic.text = quote(italic("t")),
@@ -44,7 +44,6 @@
 #'   k = 3L,
 #'   k.parameter = 3L
 #' )
-#'
 #' @export
 
 # function body